# yukka-ui/src/openApi/openApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spring Security OpenApi",
    "description": "Dokumentacja OpenApi dla Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Łukasz Domin",
      "email": "lukaszdomin@interia.pl"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "0.3"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/posty/oceny": {
      "put": {
        "tags": [
          "Post"
        ],
        "operationId": "addOcenaToPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OcenaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/pracownicy": {
      "post": {
        "tags": [
          "Pracownik"
        ],
        "operationId": "addPracownik",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "request"
                ],
                "type": "object",
                "properties": {
                  "request": {
                    "$ref": "#/components/schemas/RegistrationRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/posty": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "findAllPosty",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "szukaj",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePostResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Post"
        ],
        "operationId": "addPost",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "request"
                ],
                "type": "object",
                "properties": {
                  "request": {
                    "$ref": "#/components/schemas/PostRequest"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
...
    "/pracownicy/unban/{uzytkownik-nazwa}": {
      "patch": {
        "tags": [
          "Pracownik"
        ],
        "operationId": "unbanUzytkownik",
        "parameters": [
          {
            "name": "uzytkownik-nazwa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/pracownicy/ban": {
      "patch": {
        "tags": [
          "Pracownik"
        ],
        "operationId": "setBanUzytkownik",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "request"
                ],
                "type": "object",
                "properties": {
                  "request": {
                    "$ref": "#/components/schemas/BanRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/posty/{post-id}": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "findPostById",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Post"
        ],
        "operationId": "removePost",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/posty/{post-id}/check": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "findPostByIdCheck",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          }
        }
      }
    },
    "/posty/uzytkownik/{nazwa}": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "findAllPostyByUzytkownik",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "nazwa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePostResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/pracownicy/{uzytkownik-nazwa}": {
      "delete": {
        "tags": [
          "Pracownik"
        ],
        "operationId": "remove",
        "parameters": [
          {
            "name": "uzytkownik-nazwa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Komentarz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "komentarzId": {
            "type": "string"
          },
          "opis": {
            "type": "string"
          },
          "obraz": {
            "type": "string"
          },
          "edytowany": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time"
          },
          "rozmowaPrywatna": {
            "$ref": "#/components/schemas/RozmowaPrywatna"
          },
          "post": {
            "$ref": "#/components/schemas/Post"
          },
          "odpowiadaKomentarzowi": {
            "$ref": "#/components/schemas/Komentarz"
          },
          "odpowiedzi": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Komentarz"
            }
          },
          "uzytkownik": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "wposcie": {
            "$ref": "#/components/schemas/Post"
          },
          "ocenyLubi": {
            "type": "integer",
            "format": "int32"
          },
          "ocenyNieLubi": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Ocenil": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "lubi": {
            "type": "boolean"
          },
          "oceniany": {
            "$ref": "#/components/schemas/Oceniany"
          }
        }
      },
      "OcenilReverse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "lubi": {
            "type": "boolean"
          },
          "uzytkownik": {
            "$ref": "#/components/schemas/Uzytkownik"
          }
        }
      },
      "Ogrod": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nazwa": {
            "type": "string"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "ocenil": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcenilReverse"
            }
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "postId": {
            "type": "string"
          },
          "tytul": {
            "type": "string"
          },
          "opis": {
            "type": "string"
          },
          "obraz": {
            "type": "string"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time"
          },
          "autor": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "komentarze": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Komentarz"
            }
          },
          "komentarzeWPoscie": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Komentarz"
            }
          },
          "ocenyLubi": {
            "type": "integer",
            "format": "int32"
          },
          "ocenyNieLubi": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RozmowaPrywatna": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "aktywna": {
            "type": "boolean"
          },
          "nadawca": {
            "type": "string"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time"
          },
          "ostatnioAktualizowana": {
            "type": "string",
            "format": "date-time"
          },
          "uzytkownicy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Uzytkownik"
            }
          },
          "wiadomosci": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Komentarz"
            }
          }
        }
      },
      "Ustawienia": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "statystykiProfilu": {
            "type": "boolean"
          },
          "ogrodPokaz": {
            "type": "boolean"
          },
          "powiadomieniaKomentarzeOdpowiedz": {
            "type": "boolean"
          },
          "powiadomieniaOgrodKwitnienie": {
            "type": "boolean"
          },
          "powiadomieniaOgrodOwocowanie": {
            "type": "boolean"
          }
        }
      },
      "Uzytkownik": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "uzytId": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nazwa": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "haslo": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time"
          },
          "aktywowany": {
            "type": "boolean"
          },
          "ban": {
            "type": "boolean"
          },
          "banDo": {
            "type": "string",
            "format": "date"
          },
          "imie": {
            "type": "string"
          },
          "miasto": {
            "type": "string"
          },
          "miejsceZamieszkania": {
            "type": "string"
          },
          "opis": {
            "type": "string"
          },
          "posty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "komentarze": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Komentarz"
            }
          },
          "oceny": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ocenil"
            }
          },
          "ustawienia": {
            "$ref": "#/components/schemas/Ustawienia"
          },
          "ogrod": {
            "$ref": "#/components/schemas/Ogrod"
          },
          "rozmowyPrywatne": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RozmowaPrywatna"
            }
          },
          "blokowaniUzytkownicy": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Uzytkownik"
            }
          },
          "blokujacyUzytkownicy": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Uzytkownik"
            }
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          },
          "pracownik": {
            "type": "boolean"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "normalUzytkownik": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "accountNonLocked": {
            "type": "boolean"
          }
        }
      },
...
      "OcenaRequest": {
        "required": [
          "lubi",
          "ocenialnyId"
        ],
        "type": "object",
        "properties": {
          "ocenialnyId": {
            "type": "string"
          },
          "lubi": {
            "type": "boolean"
          }
        }
      },
      "KomentarzResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "komentarzId": {
            "type": "string"
          },
          "opis": {
            "type": "string"
          },
          "edytowany": {
            "type": "boolean"
          },
          "ocenyLubi": {
            "type": "integer",
            "format": "int32"
          },
          "ocenyNieLubi": {
            "type": "integer",
            "format": "int32"
          },
          "dataUtworzenia": {
            "type": "string"
          },
          "post": {
            "$ref": "#/components/schemas/PostResponse"
          },
          "odpowiadaKomentarzowi": {
            "$ref": "#/components/schemas/KomentarzResponse"
          },
          "odpowiedzi": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KomentarzResponse"
            }
          },
          "uzytkownikNazwa": {
            "type": "string"
          },
          "obraz": {
            "type": "string",
            "format": "byte"
          },
          "avatar": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "PostResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "postId": {
            "type": "string"
          },
          "tytul": {
            "type": "string"
          },
          "opis": {
            "type": "string"
          },
          "ocenyLubi": {
            "type": "integer",
            "format": "int32"
          },
          "ocenyNieLubi": {
            "type": "integer",
            "format": "int32"
          },
          "liczbaKomentarzy": {
            "type": "integer",
            "format": "int32"
          },
          "komentarze": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KomentarzResponse"
            }
          },
          "uzytkownik": {
            "type": "string"
          },
          "obraz": {
            "type": "string",
            "format": "byte"
          },
          "avatar": {
            "type": "string",
            "format": "byte"
          },
          "dataUtworzenia": {
            "type": "string"
          }
        }
      },
...
      "RegistrationRequest": {
        "required": [
          "email",
          "haslo",
          "nazwa",
          "powtorzHaslo"
        ],
        "type": "object",
        "properties": {
          "nazwa": {
            "maxLength": 100,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_ąćęłńóśźżĄĆĘŁŃÓŚŹŻ]+( [a-zA-Z0-9_ąćęłńóśźżĄĆĘŁŃÓŚŹŻ]+)*$",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "haslo": {
            "maxLength": 100,
            "minLength": 8,
            "type": "string"
          },
          "powtorzHaslo": {
            "maxLength": 100,
            "minLength": 8,
            "type": "string"
          }
        }
      },
...
      "PostRequest": {
        "required": [
          "tytul"
        ],
        "type": "object",
        "properties": {
          "tytul": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "opis": {
            "maxLength": 3000,
            "minLength": 0,
            "type": "string"
          },
          "obraz": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "BanRequest": {
        "required": [
          "ban",
          "banDo",
          "nazwa",
          "powod"
        ],
        "type": "object",
        "properties": {
          "nazwa": {
            "type": "string"
          },
          "powod": {
            "type": "string"
          },
          "ban": {
            "type": "boolean"
          },
          "banDo": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "PageResponsePostResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Authorization Bearer JWT",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}